# www/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Financial API",
        "description": "Wellcome! <br>To make requests, use login system passing cfpCnpj and password to get the token and click in Authorize and set the token",
        "contact": {
            "name": "Financial Team",
            "email": "gustamms@hotmail.com"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/api/type-person": {
            "get": {
                "tags": [
                    "Type person"
                ],
                "summary": "List of type person",
                "responses": {
                    "200": {
                        "description": "Returns data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "typ_id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "typ_description": {
                                                        "type": "string",
                                                        "example": "Física"
                                                    },
                                                    "typ_sisid": {
                                                        "type": "string",
                                                        "example": "FIS"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "example": "2021-05-17T22:39:57.000000Z"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "example": "2021-05-17T22:39:57.000000Z"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Error message",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object",
                                    "example": {
                                        "message": "Unauthorized"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Type person"
                ],
                "summary": "Create new type transactions",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "typ_description": {
                                        "type": "string",
                                        "example": "Lojista"
                                    },
                                    "typ_sisid": {
                                        "type": "string",
                                        "example": "LOG"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Returns data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "typ_id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "typ_description": {
                                                        "type": "string",
                                                        "example": "Física"
                                                    },
                                                    "typ_sisid": {
                                                        "type": "string",
                                                        "example": "FIS"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "example": "2021-05-17T22:39:57.000000Z"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "example": "2021-05-17T22:39:57.000000Z"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Error message",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object",
                                    "example": {
                                        "message": "Unauthorized"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/type-person/{id}": {
            "get": {
                "tags": [
                    "Type person"
                ],
                "summary": "Specific of type transactions",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Type transactions id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "typ_id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "typ_description": {
                                                        "type": "string",
                                                        "example": "Física"
                                                    },
                                                    "typ_sisid": {
                                                        "type": "string",
                                                        "example": "FIS"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "example": "2021-05-17T22:39:57.000000Z"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "example": "2021-05-17T22:39:57.000000Z"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Error message",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object",
                                    "example": {
                                        "message": "Unauthorized"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Login with CPF/CNPJ and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}